% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fish2omop.R
\name{fish2omop}
\alias{fish2omop}
\title{map fish myeloma abnormalities to OMOP from cytogenetic FISH data}
\usage{
fish2omop(df)
}
\arguments{
\item{df}{data frame with columns containing FISH abnormalities}
}
\description{
map fish myeloma abnormalities to OMOP from cytogenetic FISH data
}
\examples{
#
}
