% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fish4.R
\name{fish4all}
\alias{fish4all}
\title{extract all abnormalities from cytogenetic FISH data}
\usage{
fish4all(df, column = "region_gene")
}
\arguments{
\item{df}{data frame with a column containing FISH strings}

\item{column}{name of column containing FISH strings, default="region_gene"}
}
\description{
extract all abnormalities from cytogenetic FISH data
}
\examples{
# make some simple test data
fish_raw <- dplyr::tibble(test_id = c(1:5),
                            fishstring=c("17p13(TP53x3) CKS1Bx3",
                                      "17p13(TP53x2) CKS1Bx2",
                                      "17p13(TP53x1) CKS1Bx1 FGFR3 con IGH",
                                      "IGH con MAF",
                                      "CCND1 con IGH"))
fish_processed <- fish_raw |> fish4all("fishstring")
}
