% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fish4.R
\name{fish4cks1b}
\alias{fish4cks1b}
\title{extract cks1b abnormalities from cytogenetic FISH data}
\usage{
fish4cks1b(df, column = "region_gene")
}
\arguments{
\item{df}{data frame with a column containing FISH strings}

\item{column}{name of column containing FISH strings, default="region_gene"}
}
\description{
extract cks1b abnormalities from cytogenetic FISH data
}
\examples{
df1 <- dplyr::tibble(fishstring=c("CKS1Bx3","CKS1Bx2","CKS1Bx1"))
df2 <- df1 |> fish4cks1b(column="fishstring")
}
