% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fish4.R
\name{fish4del17p}
\alias{fish4del17p}
\title{extract del17p abnormalities from cytogenetic FISH data}
\usage{
fish4del17p(df, column = "region_gene")
}
\arguments{
\item{df}{data frame with a column containing FISH strings}

\item{column}{name of column containing FISH strings, default="region_gene"}
}
\description{
extract del17p abnormalities from cytogenetic FISH data
}
\examples{
df1 <- dplyr::tibble(fishstring=c("17p13(TP53x1)","17p13(TP53x2)","17p13(TP53x3)"))
df2 <- df1 |> fish4del17p(column="fishstring")
}
