% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fish4.R
\name{fish4ighccnd1}
\alias{fish4ighccnd1}
\title{extract ighccnd1 abnormalities from cytogenetic FISH data}
\usage{
fish4ighccnd1(df, column = "region_gene")
}
\arguments{
\item{df}{data frame with a column containing FISH strings}

\item{column}{name of column containing FISH strings, default="region_gene"}
}
\description{
extract ighccnd1 abnormalities from cytogenetic FISH data
}
\examples{
df1 <- dplyr::tibble(fishstring=c("CCND1 con IGH","IGH con CCND1","anything else"))
df2 <- df1 |> fish4ighccnd1(column="fishstring")
}
