% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fish4.R
\name{fish4ighfgfr3}
\alias{fish4ighfgfr3}
\title{extract ighfgfr3 abnormalities from cytogenetic FISH data}
\usage{
fish4ighfgfr3(df, column = "region_gene")
}
\arguments{
\item{df}{data frame with a column containing FISH strings}

\item{column}{name of column containing FISH strings, default="region_gene"}
}
\description{
extract ighfgfr3 abnormalities from cytogenetic FISH data
}
\examples{
df1 <- dplyr::tibble(fishstring=c("FGFR3 con IGH","anything else!"))
df2 <- df1 |> fish4ighfgfr3(column="fishstring")
}
