% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fish4.R
\name{fish4ighmaf}
\alias{fish4ighmaf}
\title{extract ighmaf abnormalities from cytogenetic FISH data}
\usage{
fish4ighmaf(df, column = "region_gene")
}
\arguments{
\item{df}{data frame with a column containing FISH strings}

\item{column}{name of column containing FISH strings, default="region_gene"}
}
\description{
extract ighmaf abnormalities from cytogenetic FISH data
}
\examples{
df1 <- dplyr::tibble(fishstring=c("IGH con IGH","anything else!"))
df2 <- df1 |> fish4ighmaf(column="fishstring")
}
