% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fish4myeloma-package.R
\docType{package}
\name{fish4myeloma-package}
\alias{fish4myeloma}
\alias{fish4myeloma-package}
\title{fish4myeloma: Extract Myeloma information from cytogenetic FISH data}
\description{
Extracts Myeloma abnormalities in cytogenetic FISH (Fluorescence in situ hybridization) strings.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/andysouth/fish4myeloma}
  \item Report bugs at \url{https://github.com/andysouth/fish4myeloma/issues}
}

}
\author{
\strong{Maintainer}: Andy South \email{southandy@gmail.com} (\href{https://orcid.org/0000-0003-4051-6135}{ORCID})

}
\keyword{internal}
